Update custom example with previous share extension UI
Added an additional fix for "startsWith" which may error if mimeType is
undefined

This may not be fully functional yet, based on experience with copying
the prior example app UI for the basic app. Would still need to test it
out more. But sharing a photo to the main app worked.